requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
fastapi==0.68.0
h11==0.12.0
idna==3.2
motor==2.5.1
multidict==5.1.0
pydantic==1.8.2
pymongo==3.12.0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
starlette==0.14.2
tornado==6.1
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
uvicorn==0.15.0
yarl==1.6.3
